Batching messages in pending list
Changed the way the messages are sent, always batch everything in the
pending list to submit a single remote method invocation to each client.
Also moved the reading of serverId to main (outside the server methods)
so the server gets constructed with the serverId passed as argument.
Made serverid == 0 use the default rmiregistry port (1099) for
single-server tests.