Simplified joining protocol
Changed the joining protocol to avoid having a "ready" message and
further separate server/room roles.

Now the server is only a directory of chat rooms, so client contacts it
to get the room reference and that's it.

The "connect" API was renamed "join" and moved to the chat room, so now
the client talks to the room for joining. Because of this the "ready"
API is now unnecessary, as well as some sync blocks and the timeout
mechanism for clients who connected but didn't set the "ready" flag