ClientModel harness and refactoring of names
Separated a model from the CLI client so the GUI client can use the same
code to communicate with the server and room. Now the CLI client is just
a view and the harness is the model (like a MVC architecture), the GUI
will act as another view.
I also refactored a lot of the classes names to show this MVC approach
in the naming conventions and to reduce the repetitive "Chat" prefix.